Refactoring: class GeoDataPointmap::AddInterpolatePointsCommand moved to independent files.
